Updated game map

+++ RPG/game.py
@@ -1,5 +1,6 @@
 import random
 import time
+import sys
 from items import Item
 from player import Player
 from monster import Monster
@@ -11,7 +12,12 @@ def flush():
 	print()
 	print('-'*10)
 
-# def typewrite(s):
+def typewrite(s):
+	for i in s:
+		print(i,end='')
+		sys.stdout.flush() # flush the output stream
+		time.sleep(0.03)
+	print()
 
 class Game:
 	def __init__(self):
@@ -19,31 +25,34 @@ class Game:
 		self.run = 1
 		self.damage_to = 0
 		self.damage_from = 0
+		self.traps = [(2,5)]
 		self.battle_pos = {
 		(2,1) : "Arachnid",
-		(2,2) : "Basilisk",
-		(4,3) : "Minotaur"
+		(2,3) : "Basilisk",
+		(4,4) : "Minotaur"
 		}
-		self.game_map = { # stores all the map descriptions 
+		self.game_map = { # stores all the map descriptions in a dictionary, with coordinates as the key
 		(0,0) : "The gates to the Labyrinth looms infront of you. Torches illuminate it on either side.",
 		(0,1) : "A dim glow emits from the passage ahead. The passage to the east and west are dark.",
 		(0,2) : ["A towering figure appears in front of you. [x]","The room appears to be empty."],
 		(-1,1): ["There appears to be a wooden chest lying on the ground. [x]","The chest is empty."],
 		(1,1) : "You have entered a narrow passage way. The path to the east looks menacing...",
-		(2,3) : "The roof begins to cave in... Something doesn\'t feel right. There are paths to the north and east.",
+		(2,4) : "The roof begins to cave in... Something doesn\'t feel right. There are paths to the north and east.",
 		(2,1) : ["You come face to face with an Arachnid.","The grotesque remains of the demon lies still. A passage way suddenly appears to the North."],
-		(2,4) : "You fall into a pit of lava. You die. [ENTER] to continue.",
-		(2,2) : ["You come face to face with a Basilisk.","The grotesque remains of the serpent lies still. The passage to the north extennds."],
-		(3,3) :	"Something is wrong. You feel as if you are constantly being watched... it isn't too late to turn back.",
-		(4,3) : ["Menacing...\nYou come face to face with a Minotaur","The reamins of that wretched creature is scattered around."]
+		(2,5) : "You fall into a pit of lava. You die. [ENTER] to continue.",
+		(2,3) : ["You come face to face with a Basilisk.","The grotesque remains of the serpent lies still. The passage to the north extennds."],
+		(3,4) :	"Something is wrong. You feel as if you are constantly being watched... it isn't too late to turn back. Continue on your journey East or- Oh, a passage to the north!",
+		(4,4) : ["Menacing...\nYou come face to face with a Minotaur","The reamins of that wretched creature is scattered around."],
+		(3,5) : "This, is a dead end.",
+		(2,2) : "You sense something... something slithering in the walls of the chamber... the path leads on North"
 		}
 
 	def intro(self):
-		print("Welcome to the Labyrinth, {}.".format(self.player.name))
-		# time.sleep(0)
-		print("Find your way through the maze, and defeat the mighty Minotaur.")
-		# time.sleep(2.5)
-		print("Good luck...\n")
+		typewrite("Welcome to the Labyrinth, {}.".format(self.player.name))
+		time.sleep(0.5)
+		typewrite("Find your way through the maze, and defeat the mighty Minotaur.")
+		time.sleep(1)
+		typewrite("Good luck...\n")
 
 	def end_screen(self):
 		print("Good work, {}. You have successfully defeated the Minotaur.\n".format(self.player.name))
@@ -54,8 +63,7 @@ class Game:
 	def help(self):
 		print("Actions: ")
 		print("n - move North\ne - move East\nw - move West\ns - move South")
-		print("x - interact\nh - help\ni - inventory\npick - Pick up item\nuse <item name> - Use an item")
-		print("Combat: ")
+		print("x - interact\nh - help\ni - inventory\npick - Pick up item\nuse <item name> - Use an item (only consumables)")
 
 	def show_description(self):
 		if (self.player.location == (0,2)): 
@@ -76,8 +84,8 @@ class Game:
 		print()
 		if (self.player.location == (0,2)):
 			if (not self.player.collected[0]):
-				print("A booming voice reverberates around the room.")
-				print("\"It\'s dangerous to go alone! Take this.\"")
+				typewrite("A booming voice reverberates around the room.")
+				typewrite("\"It\'s dangerous to go alone! Take this.\"")
 				item = Item("sword",1)
 				self.player.collected[0] = 1
 				self.player.add_item(item.id)	
@@ -98,12 +106,15 @@ class Game:
 		Player damage = random(attack*weapon*0.8, attack*weapon*1.2)
 		Options for player in battle:
 		1. Attack 2. Block 3. Run
+
+		Really inefficient and poorly written code... I should've planned the 
+		logic flow of the engine instead of blindly implementing it
 		'''
 		valid = ['1','2','3']
 		victor = ''
 		turn = -1
 		monster = Monster(name)
-		first = 1 if self.player.speed > monster.speed else 0
+		[first = 1 if self.player.speed > monster.speed else 0]
 		while (self.player.battle):
 			turn += 1
 			flush()
@@ -124,6 +135,7 @@ class Game:
 						print("You live to fight another day...")
 						self.player.battle = 0
 						self.player.location = self.player.prev # sets player location to where it came from
+						flush()
 						break
 			if (not self.player.battle): 
 				self.show_description()
@@ -181,8 +193,12 @@ class Game:
 					player_damage = self.player.get_damage()
 				if (monster_move =='b'):
 					player_damage //= 3
-				monster.show_message(0) if (monster_move=='a') else monster.show_message(1)
-				if (monster_move=='a'): print("The {}\'s attack deals {} damage!".format(name,monster_damage))	
+				monster.show_message(0) 
+				if (monster_move=='a'):
+					monster.show_message(0)
+					print("The {}\'s attack deals {} damage!".format(name,monster_damage))	
+				else:
+					monster.show_message(1)
 				self.player.hp -= monster_damage
 				self.damage_from += monster_damage
 				if (self.player.hp <= 0):
@@ -197,13 +213,14 @@ class Game:
 				if (monster.hp <= 0):
 					victor = 'p'
 					self.player.battle = 0  
-
+		self.player.cooldown = 0 # reset block cooldown after every battle
 		if (victor == 'p'):
 			print("You have slained the {}. [ENTER] to continue\n".format(name))
+			self.player.battle_history[self.player.location] = 1
+			if (name=="Minotaur"): return
 			drops = monster.drop()
 			for i in drops:
 				self.player.add_item(i)
-			self.player.battle_history[self.player.location] = 1
 		elif (victor == 'm'):
 			print("You have been slained by the {}... [ENTER] to continue".format(name))
 			self.player.die()
@@ -213,13 +230,13 @@ class Game:
 		self.help()
 		while (self.run):
 			flush()
-			if (self.player.battle_history[(4,3)]==1):
+			if (self.player.battle_history[(4,4)]==1):
 				self.end_screen()
 				break
 			self.show_description()
 			if (self.player.battle):
 				self.battle_engine(self.battle_pos[self.player.location])
-			if (self.player.location==(2,4)):
+			if (self.player.location in self.traps):
 				self.player.die()
 			action = input('>')
 			if (action.lower() == 'n'):
